Fixed Croak servie

+++ backend/src/main/java/com/raven/croaker/rest/CroakController.java
@@ -1,12 +1,19 @@
 package com.raven.croaker.rest;
 
 import com.raven.croaker.domain.Croak;
+import com.raven.croaker.domain.User;
 import com.raven.croaker.dto.CroakDTO;
 import com.raven.croaker.service.CroakService;
+import com.raven.croaker.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+import java.security.Principal;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/api/croak")
@@ -14,25 +21,45 @@ public class CroakController {
     @Autowired
     private CroakService croakService;
 
+    @Autowired
+    private UserService userService;
+
     @ResponseBody
-    @RequestMapping(value = "/", method = RequestMethod.GET)
+    @RequestMapping(method = RequestMethod.GET)
     public ResponseEntity<Iterable<Croak>> getAll() {
         return new ResponseEntity<>(croakService.findAll(), HttpStatus.OK);
     }
 
     @ResponseBody
-    @RequestMapping(value = "/", method = RequestMethod.POST)
-    public ResponseEntity<Croak> saveCroak(@RequestBody CroakDTO croakDTO) {
+    @RequestMapping(value = "/{username}", method = RequestMethod.GET)
+    public ResponseEntity<Iterable<Croak>> getAllFromUser(@PathVariable("username") String username) {
+        return new ResponseEntity<>(croakService.findAllFromUser(username), HttpStatus.OK);
+    }
+
+    @ResponseBody
+    @RequestMapping(method = RequestMethod.POST)
+    @PreAuthorize("hasAuthority('STANDARD_USER')")
+    public ResponseEntity<Croak> saveCroak(@RequestBody CroakDTO croakDTO, HttpServletRequest request) {
+        Principal principal = request.getUserPrincipal();
+        User user = userService.findByUsername(principal.getName());
         Croak croak = new Croak();
         croak.setMessage(croakDTO.getMessage());
+        croak.setUserId(user.getId());
+        croak.setUsername(user.getUsername());
         return new ResponseEntity<>(croakService.save(croak), HttpStatus.CREATED);
     }
 
     @ResponseBody
-    @RequestMapping(value = "/", method = RequestMethod.DELETE)
-    public ResponseEntity<Croak> deleteCroak(@RequestBody Croak croak) {
-        croakService.delete(croak);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    @RequestMapping(method = RequestMethod.DELETE)
+    @PreAuthorize("hasAuthority('STANDARD_USER')")
+    public ResponseEntity<Croak> deleteCroak(@RequestBody Croak croak, HttpServletRequest request) {
+        Principal principal = request.getUserPrincipal();
+        User user = userService.findByUsername(principal.getName());
+        if(croak.getUserId().equals(user.getId()) && Objects.equals(croak.getUsername(), user.getUsername())) {
+            croakService.delete(croak);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
 }
